refactor(api): renamed files without folder's name for consistency/simplicity
