+++ api/src/utils/helpers/isPasswordValid.ts
